+++ src/gatelin-express.d.ts
@@ -1,7 +1,7 @@
 import type { Request, Response, NextFunction } from 'express';
 
 interface Consumer {
-  id: number;
+  userId: number;
   nickname: string;
 }